Update nodejs.sh
Redid install code for nvm due to an issue on debian where the shell was refusing to execute source ~/.bashrc which is needed for the shell to see the newly added nvm path causing it to not be found

+++ modules/other/development/nodejs.sh
@@ -5,9 +5,16 @@
 #LTS release.
 
 install_nodejs(){
-	curl -o- https://raw.githubusercontent.com/nvm-sh/nvm/v0.40.0/install.sh | bash
-	source ~/.bashrc
-	nvm install lts/*
-}
+    test -d "$HOME"/.nvm && NVM_FOLDER="exists"
+    if [ "$NVM_FOLDER" = "exists" ];
+        then
+           nvm install lts/*
+    elif [ "$NVM_FOLDER" = "missing" ];
+        then
+			curl -o- https://raw.githubusercontent.com/nvm-sh/nvm/v0.40.0/install.sh | bash
+			zenity --info --text="Please rerun this option after restarting your shell otherwise it won't see nvm"
 
+    fi
+}
+NVM_FOLDER="missing"
 install_nodejs
